% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kansler grafik.R
\name{gguka_bar}
\alias{gguka_bar}
\title{UKA stapeldiagram 1}
\usage{
gguka_bar(df, x = x, y = y, g = g, pal = "uka_1", format = Svensk_antal)
}
\arguments{
\item{df}{dataframe}

\item{x}{x-variabel}

\item{y}{y-variabel (num)}

\item{g}{gruppvariabel}

\item{format}{format. default Svensk antal}

\item{t}{fargpalett (se lista i uka_paletter)}
}
\value{
stapeldiagram (staplat)
}
\description{
använder theme_uka1
}
\examples{
uka_line(df1 = df, x = nam, y = num, g = gro)

}
