% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kansler grafik.R
\name{gguka_bar1}
\alias{gguka_bar1}
\title{UKA stapeldiagram 1}
\usage{
gguka_bar1(df1, x, y, g, pal)
}
\arguments{
\item{x}{x-variabel}

\item{y}{y-variabel (num)}

\item{g}{gruppvariabel}

\item{t}{fargpalett (se lista i uka_farg_paletter)}

\item{df}{dataframe}
}
\value{
stapeldiagram (staplat)
}
\description{
använder theme_uka1
}
\examples{
uka_line(df1 = df, x = nam, y = num, g = gro)

}
