% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kansler grafik.R
\name{gguka_line}
\alias{gguka_line}
\title{UKA linjediagram 1}
\usage{
gguka_line(df, x = x, y = y, g = g, pal = "uka_1", format = Svensk_antal)
}
\arguments{
\item{df}{dataframe}

\item{x}{x-variabel}

\item{y}{y-variabel (num)}

\item{g}{gruppvariabel}

\item{t}{fargpalett (se lista i uka_farg_paletter)}

\item{format.}{default Svensk antal}
}
\value{
linjediagram
}
\description{
använder theme_uka1#'
}
