% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard.R
\docType{data}
\name{uka_farg_vect}
\alias{uka_farg_vect}
\title{Vektor med kanslerns farger}
\format{
An object of class \code{character} of length 18.
}
\usage{
uka_farg_vect
}
\description{
18 farger:
nr. 1-3 uka:s huvudsakliga profilfarger
nr. 4-6 sekundarfarger
}
\keyword{datasets}
