% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard.R
\name{gguka_line}
\alias{gguka_line}
\title{UKA linjediagram 1}
\usage{
gguka_line(df1, x, y, g)
}
\arguments{
\item{x}{x-variabel}

\item{y}{y-variabel (num)}

\item{g}{gruppvariabel}

\item{df}{dataframe}
}
\value{
linjediagram
}
\description{
använder theme_uka1 och uka_farg. for testning
}
\examples{
anvander make_df: uka_line(df1 = df, x = nam, y = num, g = gro)

}
