% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard.R
\docType{data}
\name{uka_farg_vect}
\alias{uka_farg_vect}
\title{Vektor med grundfarger}
\format{
An object of class \code{character} of length 12.
}
\usage{
uka_farg_vect
}
\description{
12 farger, varav de tre forsta ar uka:s huvudsakliga profilfarger. resten
ar hamtade fran arsrapporten 2020 samt pahittade
}
\keyword{datasets}
