% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kansler grafik.R
\name{uka_tema}
\alias{uka_tema}
\title{Tema UKA}
\usage{
uka_tema(size = 10, angle = "a0")
}
\arguments{
\item{size}{anvands for att ange storlek pa text. default = 10}

\item{angle}{ange vinkel pa x-axelns text. "a0" default. "a90" ger 90 grader.}
}
\description{
Tema som anvands for formatering av figurer. Ar tankt att designad for uttag i formatet 4:6 / landscape.
}
